+++ src/main/java/cn/xidian/meetingroom/service/LogService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.time.LocalDateTime;
 
 public interface LogService {
-    LogWithBLOBs getLogById(Integer logId);
+    LogWithBLOBs getLogById(Long logId);
     List<LogWithBLOBs> getLogsByUserId(Integer userId);
     List<LogWithBLOBs> getLogsByDateRange(LocalDateTime startDate, LocalDateTime endDate);
     List<LogWithBLOBs> getLogsByOperationType(String operationType);
