+++ src/main/java/cn/xidian/meetingroom/service/ReservationService.java
@@ -6,13 +6,13 @@ import java.util.List;
 import java.time.LocalDateTime;
 
 public interface ReservationService {
-    ReservationWithBLOBs getReservationById(Integer reservationId);
+    ReservationWithBLOBs getReservationById(Long reservationId);
     List<ReservationWithBLOBs> getReservationsByUserId(Integer userId);
     List<ReservationWithBLOBs> getReservationsByMeetingRoomId(Integer meetingRoomId);
     List<ReservationWithBLOBs> getReservationsByTimeRange(Integer meetingRoomId, LocalDateTime startTime, LocalDateTime endTime);
     ReservationWithBLOBs createReservation(ReservationWithBLOBs reservation);
-    ReservationWithBLOBs updateReservation(Integer reservationId, ReservationWithBLOBs reservation);
-    void deleteReservation(Integer reservationId);
-    ReservationWithBLOBs approveReservation(Integer reservationId);
-    ReservationWithBLOBs rejectReservation(Integer reservationId, String rejectionReason);
+    ReservationWithBLOBs updateReservation(Long reservationId, ReservationWithBLOBs reservation);
+    void deleteReservation(Long reservationId);
+    ReservationWithBLOBs approveReservation(Long reservationId);
+    ReservationWithBLOBs rejectReservation(Long reservationId, String rejectionReason);
 }
